+++ packages/dw-react-controls/src/components/AccessibleTreeItem/index.ts
@@ -1,0 +1,3 @@
+export * from "./AccessibleTreeItem";
+export * from "./AccessibleTreeItem.base";
+export * from "./AccessibleTreeItem.types";
