+++ packages/dw-react-controls/src/components/AccessibleTreeItem/useTreeItem.ts
@@ -1,5 +1,5 @@
 import * as React from "react";
-import TreeView2Context from "../TreeView2/TreeView2.context";
+import AccessibleTreeViewContext from "../AccessibleTreeView/AccessibleTreeView.context";
 
 export default function useTreeItem(nodeId: string) {
 	const {
@@ -12,7 +12,7 @@ export default function useTreeItem(nodeId: string) {
 		multiSelect,
 		selectNode,
 		toggleExpansion
-	} = React.useContext(TreeView2Context);
+	} = React.useContext(AccessibleTreeViewContext);
 
 	const disabled = isDisabled ? isDisabled(nodeId) : false;
 	const expandable = isExpandable ? isExpandable(nodeId) : false;
