+++ packages/dw-react-controls/src/components/TaxonomyPicker/TaxonomyPickerDialog/TaxonomyPickerDialog.base.tsx
@@ -4,9 +4,10 @@ import { DialogFooter } from "office-ui-fabric-react/lib/Dialog";
 import { classNamesFunction } from "office-ui-fabric-react/lib/Utilities";
 import * as React from "react";
 import * as rfdc from "rfdc";
+import { AccessibleTreeItem } from "../../AccessibleTreeItem";
+import { AccessibleTreeView } from "../../AccessibleTreeView";
 import { ITermValue, TermPicker } from "../../TermPicker";
-import { ITreeItemAction, TreeItem } from "../../TreeItem";
-import { TreeView } from "../../TreeView";
+import { ITreeItemAction } from "../../TreeItem";
 import { WideDialog } from "../../WideDialog";
 import { ITerm, ITermCreationResult } from "../models";
 import { TermAdder } from "../TermAdder";
@@ -225,14 +226,14 @@ export const TaxonomyPickerDialogBase: React.FC<ITaxonomyPickerDialogProps> = (p
 		}
 
 		return (
-			<TreeView
+			<AccessibleTreeView
 				expanded={expandedNodes}
 				selected={selectedTreeItem}
 				onNodeSelect={handleTreeSelection}
 				onNodeToggle={handleNodeToggle}
 			>
 				{showRootNode ? (
-					<TreeItem
+					<AccessibleTreeItem
 						nodeId={rootNodeKey}
 						label={rootNodeLabel || ""}
 						disabled={true}
@@ -240,11 +241,11 @@ export const TaxonomyPickerDialogBase: React.FC<ITaxonomyPickerDialogProps> = (p
 						actions={treeItemActions}
 					>
 						{nodeTree}
-					</TreeItem>
+					</AccessibleTreeItem>
 				) : (
 					<>{nodeTree}</>
 				)}
-			</TreeView>
+			</AccessibleTreeView>
 		);
 	};
 
@@ -280,7 +281,7 @@ export const TaxonomyPickerDialogBase: React.FC<ITaxonomyPickerDialogProps> = (p
 		}
 
 		return (
-			<TreeItem
+			<AccessibleTreeItem
 				key={term.key}
 				nodeId={term.key}
 				label={term.name}
@@ -290,7 +291,7 @@ export const TaxonomyPickerDialogBase: React.FC<ITaxonomyPickerDialogProps> = (p
 				actions={treeItemActions}
 			>
 				{children}
-			</TreeItem>
+			</AccessibleTreeItem>
 		);
 	};
