--- packages/dw-react-controls/src/components/TreeItem2/index.ts
@@ -1,3 +1,0 @@
-export * from "./TreeItem2";
-export * from "./TreeItem2.base";
-export * from "./TreeItem2.types";
