--- packages/dw-react-controls/src/components/TreeView2/index.ts
@@ -1,5 +1,0 @@
-export * from "./TreeView2";
-export * from "./TreeView2.base";
-export * from "./TreeView2.context";
-export { default as TreeView2Context } from "./TreeView2.context";
-export * from "./TreeView2.types";
