--- packages/dw-react-controls/src/components/TreeView2/stories/TreeView2.stories.mdx
@@ -1,21 +1,0 @@
-import { ArgsTable, Meta, Story, Canvas } from "@storybook/addon-docs";
-import * as stories from "./TreeView2.stories";
-import { TreeView2 } from "../TreeView2";
-import { ITreeView2Context } from "../TreeView2.context";
-import { TreeItem2 } from "../../TreeItem2";
-
-<Meta title="Components/TreeView2" component={TreeView2} subcomponents={{ TreeItem2 }} />
-
-# TreeView2
-
-A `TreeView` represents a hierarchical list.
-
-### Basic usage
-
-<Canvas withToolbar={true}>
-	<Story story={stories.Basic} name="Basic usage" />
-</Canvas>
-
-## Properties
-
-<ArgsTable components={{ TreeView2, TreeItem2 }} />
